Refactor to FTC RobotController System of File Organization
