+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pedroPathing/examples/ExampleFieldCentricTeleop.java
@@ -1,11 +1,11 @@
-package pedroPathing.examples;
+package org.firstinspires.ftc.teamcode.pedroPathing.examples;
 
 import com.pedropathing.follower.Follower;
 import com.pedropathing.geometry.Pose;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import pedroPathing.Constants;
+import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
 
 /**
  * This is an example teleop that showcases movement and field-centric driving.
