+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pedroPathing/tuners_tests/localization/ForwardTuner.java
@@ -1,25 +1,25 @@
-package pedroPathing.tuners_tests.localization;
+package org.firstinspires.ftc.teamcode.pedroPathing.tuners_tests.localization;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.pedropathing.follower.Follower;
+import com.pedropathing.util.Drawing;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import com.pedropathing.util.DashboardPoseTracker;
-import com.pedropathing.util.Drawing;
 
-import pedroPathing.Constants;
+import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
 
 /**
- * This is the LateralTuner OpMode. This tracks the strafe movement of the robot and displays the
+ * This is the ForwardTuner OpMode. This tracks the forward movement of the robot and displays the
  * necessary ticks to inches multiplier. This displayed multiplier is what's necessary to scale the
  * robot's current distance in ticks to the specified distance in inches. So, to use this, run the
  * tuner, then pull/push the robot to the specified distance using a ruler on the ground. When you're
  * at the end of the distance, record the ticks to inches multiplier. Feel free to run multiple trials
- * and average the results. Then, input the multiplier into the strafe ticks to inches in your
+ * and average the results. Then, input the multiplier into the forward ticks to inches in your
  * localizer of choice.
  * You can adjust the target distance on FTC Dashboard: 192/168/43/1:8080/dash
  *
@@ -27,8 +27,8 @@ import pedroPathing.Constants;
  * @version 1.0, 5/6/2024
  */
 @Config
-@Autonomous(name = "Lateral Localizer Tuner", group = ".Localization")
-public class LateralTuner extends OpMode {
+@Autonomous(name = "Forward Localizer Tuner", group = ".Localization")
+public class ForwardTuner extends OpMode {
     private Follower follower;
     private DashboardPoseTracker dashboardPoseTracker;
 
@@ -46,11 +46,8 @@ public class LateralTuner extends OpMode {
         dashboardPoseTracker = follower.getDashboardPoseTracker();
 
         telemetryA = new MultipleTelemetry(this.telemetry, FtcDashboard.getInstance().getTelemetry());
-        telemetryA.addLine("Pull your robot to the right " + DISTANCE + " inches. Your strafe ticks to inches will be shown on the telemetry.");
+        telemetryA.addLine("Pull your robot forward " + DISTANCE + " inches. Your forward ticks to inches will be shown on the telemetry.");
         telemetryA.update();
-
-        Drawing.drawRobot(follower.getPose(), "#4CAF50");
-        Drawing.sendPacket();
     }
 
     /**
@@ -61,9 +58,9 @@ public class LateralTuner extends OpMode {
     public void loop() {
         follower.update();
 
-        telemetryA.addData("distance moved", follower.getPose().getY());
-        telemetryA.addLine("The multiplier will display what your strafe ticks to inches should be to scale your current distance to " + DISTANCE + " inches.");
-        telemetryA.addData("multiplier", DISTANCE / (follower.getPose().getY() / follower.getPoseTracker().getLocalizer().getLateralMultiplier()));
+        telemetryA.addData("distance moved", follower.getPose().getX());
+        telemetryA.addLine("The multiplier will display what your forward ticks to inches should be to scale your current distance to " + DISTANCE + " inches.");
+        telemetryA.addData("multiplier", DISTANCE / (follower.getPose().getX() / follower.getPoseTracker().getLocalizer().getForwardMultiplier()));
         telemetryA.update();
 
         Drawing.drawPoseHistory(dashboardPoseTracker, "#4CAF50");
