+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pedroPathing/tuners_tests/localization/TurnTuner.java
@@ -1,40 +1,38 @@
-package pedroPathing.tuners_tests.localization;
+package org.firstinspires.ftc.teamcode.pedroPathing.tuners_tests.localization;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.pedropathing.follower.Follower;
-import com.pedropathing.util.Drawing;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import com.pedropathing.util.DashboardPoseTracker;
-
-import pedroPathing.Constants;
+import com.pedropathing.util.Drawing;
+import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
 
 /**
- * This is the ForwardTuner OpMode. This tracks the forward movement of the robot and displays the
+ * This is the TurnTuner OpMode. This tracks the turning movement of the robot and displays the
  * necessary ticks to inches multiplier. This displayed multiplier is what's necessary to scale the
- * robot's current distance in ticks to the specified distance in inches. So, to use this, run the
- * tuner, then pull/push the robot to the specified distance using a ruler on the ground. When you're
- * at the end of the distance, record the ticks to inches multiplier. Feel free to run multiple trials
- * and average the results. Then, input the multiplier into the forward ticks to inches in your
- * localizer of choice.
- * You can adjust the target distance on FTC Dashboard: 192/168/43/1:8080/dash
+ * robot's current angle in ticks to the specified angle in radians. So, to use this, run the
+ * tuner, then pull/push the robot to the specified angle using a protractor or lines on the ground.
+ * When you're at the end of the angle, record the ticks to inches multiplier. Feel free to run
+ * multiple trials and average the results. Then, input the multiplier into the turning ticks to
+ * radians in your localizer of choice.
+ * You can adjust the target angle on FTC Dashboard: 192/168/43/1:8080/dash
  *
  * @author Anyi Lin - 10158 Scott's Bots
  * @version 1.0, 5/6/2024
  */
 @Config
-@Autonomous(name = "Forward Localizer Tuner", group = ".Localization")
-public class ForwardTuner extends OpMode {
+@Autonomous(name = "Turn Localizer Tuner", group = ".Localization")
+public class TurnTuner extends OpMode {
     private Follower follower;
     private DashboardPoseTracker dashboardPoseTracker;
 
     private Telemetry telemetryA;
 
-    public static double DISTANCE = 48;
+    public static double ANGLE = 2 * Math.PI;
 
     /**
      * This initializes the PoseUpdater as well as the FTC Dashboard telemetry.
@@ -46,8 +44,11 @@ public class ForwardTuner extends OpMode {
         dashboardPoseTracker = follower.getDashboardPoseTracker();
 
         telemetryA = new MultipleTelemetry(this.telemetry, FtcDashboard.getInstance().getTelemetry());
-        telemetryA.addLine("Pull your robot forward " + DISTANCE + " inches. Your forward ticks to inches will be shown on the telemetry.");
+        telemetryA.addLine("Turn your robot " + ANGLE + " radians. Your turn ticks to inches will be shown on the telemetry.");
         telemetryA.update();
+
+        Drawing.drawRobot(follower.getPose(), "#4CAF50");
+        Drawing.sendPacket();
     }
 
     /**
@@ -58,10 +59,9 @@ public class ForwardTuner extends OpMode {
     public void loop() {
         follower.update();
 
-        telemetryA.addData("distance moved", follower.getPose().getX());
-        telemetryA.addLine("The multiplier will display what your forward ticks to inches should be to scale your current distance to " + DISTANCE + " inches.");
-        telemetryA.addData("multiplier", DISTANCE / (follower.getPose().getX() / follower.getPoseTracker().getLocalizer().getForwardMultiplier()));
-        telemetryA.update();
+        telemetryA.addData("total angle", follower.getTotalHeading());
+        telemetryA.addLine("The multiplier will display what your turn ticks to inches should be to scale your current angle to " + ANGLE + " radians.");
+        telemetryA.addData("multiplier", ANGLE / (follower.getTotalHeading() / follower.getPoseTracker().getLocalizer().getTurningMultiplier()));
 
         Drawing.drawPoseHistory(dashboardPoseTracker, "#4CAF50");
         Drawing.drawRobot(follower.getPose(), "#4CAF50");
